+++ bluegraph/downstream/__init__.py
@@ -14,7 +14,5 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 from .data_structures import (ElementClassifier,
-                              Preprocessor,
-                              Embedder,
                               EmbeddingPipeline)
 from .utils import *
